Update A.06.IR15.srv.keyword.md
Test cases A.05.IR14 and A.06.IR15 do not specifically test the existence of the keyword element. I.e. if the test case fails, it is not clear whether this is because of a lack of keyword element or because of the lack of a keyword originating from the GEMET Thesaurus. Thefore the following change is proposed:

Add A.13.IR13 as a prerequisite to test cases A.05.IR14 and A.06.IR15